trying to solve header file warnings (unused static functions)
