+++ libs/ofxLibwebsockets/include/ofxLibwebsockets/Client.h
@@ -28,20 +28,7 @@ namespace ofxLibwebsockets {
     };
     
     // call this function to set up a vanilla client options object
-    static ClientOptions defaultClientOptions(){
-        ClientOptions opts;
-        opts.host     = "localhost";
-        opts.port     = 80;
-        opts.bUseSSL  = false;
-        opts.channel  = "/";
-        opts.protocol = "NULL";
-        opts.version  = -1;     //use latest version
-        
-        opts.ka_time      = 0;
-        opts.ka_probes    = 0;
-        opts.ka_interval  = 0;
-        return opts;
-    };
+    extern ClientOptions defaultClientOptions();
     
     class Client : public Reactor {
         friend class Protocol;
